Search Engine Assignment 2/requirements.txt
flask
nltk
tqdm
requests
beautifulsoup4
Pillow
regex
openai-clip
torch
transformers
scikit-learn
